# internal/webserver/swaggerdocs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for the honey web UI (`honey web`). Authenticate with the same token printed at startup, or `HONEY_WEB_TOKEN`. WebSocket endpoints `GET /ws/ssh` and `GET /ws/pve-qemu-vnc` are not described by this OpenAPI document.",
        "title": "Honey Web API",
        "contact": {
            "name": "honey"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8765",
    "basePath": "/",
    "paths": {
...
        "/api/v1/recipes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "List recipe files",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.RecipesListResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/recipes/graph-plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Build graph recipe plan",
                "parameters": [
                    {
                        "description": "path or recipe_content",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.GraphPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cuetry.RecipeGraphPlan"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recipes/parse": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Parse recipe file to JSON",
                "parameters": [
                    {
                        "description": "absolute recipe file path",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.RecipesParseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.RecipesParseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/recipes/validate-content": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recipes"
                ],
                "summary": "Validate recipe JSON",
                "parameters": [
                    {
                        "description": "recipe object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.ValidateContentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.ValidateContentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/webserver.ValidateContentResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/recordings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "List session recordings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter by provider",
                        "name": "provider",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by host name",
                        "name": "host_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by host IP",
                        "name": "host_ip",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.RecordingsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/recordings/play": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recordings"
                ],
                "summary": "Load recording events",
                "parameters": [
                    {
                        "description": "recording file name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webserver.RecordingsPlayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/webserver.RecordingsPlayResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cuetry.GraphPlanEdge": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "cuetry.GraphPlanNode": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "kv_tunnel": {
                    "type": "boolean"
                },
                "preview": {
                    "type": "string"
                },
                "wave": {
                    "type": "integer"
                },
                "when": {
                    "type": "string"
                }
            }
        },
        "cuetry.RecipeGraphPlan": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cuetry.GraphPlanEdge"
                    }
                },
                "mermaid": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cuetry.GraphPlanNode"
                    }
                },
                "type": {
                    "type": "string"
                },
                "waves": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/cuetry.GraphPlanNode"
                        }
                    }
                }
            }
        },
...
        "recordings.Event": {
            "type": "object",
            "properties": {
                "cols": {
                    "type": "integer"
                },
                "data_b64": {
                    "type": "string"
                },
                "direction": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "rows": {
                    "type": "integer"
                },
                "time_ms": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "webserver.GraphPlanRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "recipe_content": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "webserver.RecipeListEntry": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "webserver.RecipesListResponse": {
            "type": "object",
            "properties": {
                "recipes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webserver.RecipeListEntry"
                    }
                }
            }
        },
        "webserver.RecipesParseRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
        "webserver.RecipesParseResponse": {
            "type": "object",
            "properties": {
                "recipe": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "webserver.RecordingListEntry": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "host_ip": {
                    "type": "string"
                },
                "host_name": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "modified_unix_ms": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "size_bytes": {
                    "type": "integer"
                },
                "trigger": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "webserver.RecordingsListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webserver.RecordingListEntry"
                    }
                }
            }
        },
        "webserver.RecordingsPlayRequest": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                }
            }
        },
        "webserver.RecordingsPlayResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/recordings.Event"
                    }
                },
                "file_name": {
                    "type": "string"
                }
            }
        },
        "webserver.ResolvedStepSummary": {
            "type": "object",
            "properties": {
                "depends": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "preview": {
                    "type": "string"
                },
                "run_as": {
                    "type": "string"
                },
                "wave": {
                    "type": "integer"
                },
                "when": {
                    "type": "string"
                }
            }
        },
...
        "webserver.ValidateContentError": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "webserver.ValidateContentRequest": {
            "type": "object",
            "properties": {
                "recipe_content": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "webserver.ValidateContentResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webserver.ValidateContentError"
                    }
                },
                "graph": {
                    "$ref": "#/definitions/cuetry.RecipeGraphPlan"
                },
                "plan": {
                    "type": "string"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/webserver.ResolvedStepSummary"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use `Authorization: Bearer \u003ctoken\u003e` where `\u003ctoken\u003e` is the web UI token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "HoneyTokenHeader": {
            "description": "Alternative header: `X-Honey-Token: \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "X-Honey-Token",
            "in": "header"
        },
        "TokenQuery": {
            "description": "Optional query token (same value) for URLs: `?token=\u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}